Small changes to the list layout Type file splitting
