+++ src/types/MetaData.ts
@@ -1,0 +1,22 @@
+import { IPicture } from 'music-metadata-browser'
+
+export interface MetaData {
+  path: string[];
+  size?: number;
+  title: string;
+  artist?: string | undefined;
+  albumArtist?: string | undefined;
+  album?: string | undefined;
+  year?: number | undefined;
+  genre?: string[] | undefined;
+  cover?: IPicture[] | undefined;
+}
+
+export interface MetaDataListStatus {
+  metaDataList: MetaData[];
+}
+
+export interface MetaDataListAction {
+  updateMetaDataList: (metaDataList: MetaDataListStatus['metaDataList']) => void
+  insertMetaDataList: (metaData: MetaData) => void
+}
