+++ src/types/file.ts
@@ -1,0 +1,6 @@
+export interface File {
+  fileName: string;
+  filePath: string[];
+  fileSize: number;
+  fileType: 'folder' | 'audio' | 'video' | 'picture' | 'other';
+}
